# apps/backend/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/auth/signup": {
            "post": {
                "operationId": "AuthController_signup",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignupDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                },
                "summary": "Register a new user",
                "tags": [
                    "Authentication"
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "operationId": "AuthController_login",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful, returns JWT token"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "summary": "Login with email and password",
                "tags": [
                    "Authentication"
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "operationId": "AuthController_refresh",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully"
                    },
                    "401": {
                        "description": "Invalid or expired refresh token"
                    }
                },
                "summary": "Refresh JWT token",
                "tags": [
                    "Authentication"
                ]
            }
        },
...
        "/api/chapters": {
            "get": {
                "operationId": "ChaptersController_findBySession",
                "parameters": [
                    {
                        "name": "sessionId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chapters"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get chapters for session",
                "tags": [
                    "Chapters"
                ]
            },
            "post": {
                "operationId": "ChaptersController_create",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": "Chapter created"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Create new chapter",
                "tags": [
                    "Chapters"
                ]
            }
        },
        "/api/chapters/{id}": {
            "get": {
                "operationId": "ChaptersController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chapter details"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get chapter by ID",
                "tags": [
                    "Chapters"
                ]
            },
            "patch": {
                "operationId": "ChaptersController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chapter updated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Update chapter",
                "tags": [
                    "Chapters"
                ]
            },
            "delete": {
                "operationId": "ChaptersController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chapter deleted"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Delete chapter",
                "tags": [
                    "Chapters"
                ]
            }
        },
        "/api/chapter-events": {
            "get": {
                "operationId": "ChapterEventsController_find",
                "parameters": [
                    {
                        "name": "chapterId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sessionId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chapter events"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get chapter events",
                "tags": [
                    "Chapter Events"
                ]
            },
            "post": {
                "operationId": "ChapterEventsController_create",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": "Chapter event created"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Create new chapter event",
                "tags": [
                    "Chapter Events"
                ]
            }
        },
        "/api/chapter-events/{id}": {
            "get": {
                "operationId": "ChapterEventsController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chapter event details"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get chapter event by ID",
                "tags": [
                    "Chapter Events"
                ]
            },
            "patch": {
                "operationId": "ChapterEventsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chapter event updated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Update chapter event",
                "tags": [
                    "Chapter Events"
                ]
            },
            "delete": {
                "operationId": "ChapterEventsController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chapter event deleted"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Delete chapter event",
                "tags": [
                    "Chapter Events"
                ]
            }
        },
...
        "/api/ai/generate-avatar": {
            "post": {
                "description": "Uses AI to generate a character portrait based on physical description and attributes",
                "operationId": "AiController_generateAvatar",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenerateAvatarDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "summary": "Generate an AI avatar image for a character",
                "tags": [
                    "ai"
                ]
            }
        },
        "/api/ai/generate-scene": {
            "post": {
                "description": "Creates an atmospheric scene image based on description, location, and mood",
                "operationId": "AiController_generateScene",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenerateSceneDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "summary": "Generate an AI scene image",
                "tags": [
                    "ai"
                ]
            }
        },
        "/api/ai/suggest-narrative": {
            "post": {
                "description": "Generates narrative suggestions based on current game context and recent events",
                "operationId": "AiController_suggestNarrative",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SuggestNarrativeDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "summary": "Get AI narrative suggestions",
                "tags": [
                    "ai"
                ]
            }
        },
...
    },
    "info": {
        "title": "Game-Plug API",
        "description": "Call of Cthulhu 7e RPG Platform API - Complete backend specification with NestJS and OpenAPI 3.0",
        "version": "1.0.0",
        "contact": {
            "name": "Game-Plug Team",
            "url": "https://github.com/robinswood",
            "email": "contact@game-plug.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
...
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "Development"
        },
        {
            "url": "http://localhost:4000",
            "description": "Production"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "SignupDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@example.com",
                        "description": "User email address (must be unique)"
                    },
                    "password": {
                        "type": "string",
                        "example": "SecurePassword123!",
                        "description": "Password (8-128 chars, 1 uppercase, 1 digit, 1 special char)"
                    },
                    "firstName": {
                        "type": "string",
                        "example": "Jean",
                        "description": "First name (2-50 characters)"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Dupont",
                        "description": "Last name (2-50 characters)"
                    }
                },
                "required": [
                    "email",
                    "password",
                    "firstName",
                    "lastName"
                ]
            },
            "LoginDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@example.com",
                        "description": "User email address"
                    },
                    "password": {
                        "type": "string",
                        "example": "SecurePassword123!",
                        "description": "User password"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
...
            "GenerateAvatarDto": {
                "type": "object",
                "properties": {
                    "characterName": {
                        "type": "string",
                        "description": "Character name"
                    },
                    "occupation": {
                        "type": "string",
                        "description": "Character occupation"
                    },
                    "physicalDescription": {
                        "type": "string",
                        "description": "Physical description (height, build, hair, eyes, etc.)"
                    },
                    "age": {
                        "type": "string",
                        "description": "Character age"
                    },
                    "gender": {
                        "type": "string",
                        "description": "Character gender"
                    },
                    "styleHints": {
                        "type": "string",
                        "description": "Additional style instructions for the AI"
                    }
                },
                "required": [
                    "characterName",
                    "occupation"
                ]
            },
            "GenerateSceneDto": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Scene title or theme"
                    },
                    "description": {
                        "type": "string",
                        "description": "Scene description or context"
                    },
                    "location": {
                        "type": "string",
                        "description": "Location of the scene"
                    },
                    "timePeriod": {
                        "type": "string",
                        "description": "Time period (e.g., 1920s, modern, Victorian)"
                    },
                    "mood": {
                        "type": "string",
                        "description": "Mood or atmosphere (e.g., tense, mysterious, calm)"
                    },
                    "characters": {
                        "description": "Characters present in the scene",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "title",
                    "description"
                ]
            },
            "SuggestNarrativeDto": {
                "type": "object",
                "properties": {
                    "sessionId": {
                        "type": "string",
                        "description": "Session ID for context"
                    },
                    "context": {
                        "type": "string",
                        "description": "Current situation or context"
                    },
                    "recentEvents": {
                        "description": "Recent events or actions taken",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "characters": {
                        "description": "Character names involved",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "tone": {
                        "type": "string",
                        "description": "Desired tone (horror, mystery, action, etc.)"
                    },
                    "suggestionCount": {
                        "type": "string",
                        "description": "Number of suggestions to generate"
                    }
                },
                "required": [
                    "sessionId",
                    "context"
                ]
            },
...
        }
    }
}